+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Added `MLBMatchup` model and `MatchupRunner` to scrape MLB matchups from https://baseball-reference.com
 - Added EventID to sportsreference util and updated `MatchupRunner.GetMatchups(...)` to use it.
 ### Changed
 - Moved local extractPlayerID to sportsreference util and renamed it PlayerID